rename source-date-axis.ts to source-date-axis.component.ts
