+++ src/app/subject/components/source-graphs/source-date-axis/source-date-axis.component.scss
