+++ src/app/subject/components/source-graphs/source-graphs.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core'
 
 import { ChartsModule } from '../../../charts/charts.module'
 import { MaterialModule } from '../../../material'
-import { SourceDateAxisComponent } from './source-date-axis/source-date-axis'
+import { SourceDateAxisComponent } from './source-date-axis/source-date-axis.component'
 import { SourceGraphComponent } from './source-graph/source-graph.component'
 import { SourceGraphsComponent } from './source-graphs.component'
 import { SourceTooltipComponent } from './source-tooltip/source-tooltip.component'
